+++ spec/lib/00_sqlite_aware_spec.rb
@@ -8,6 +8,7 @@ require_relative 'filter_model/filter_map_block_spec'
 require_relative 'filter_model/affix_synonyms_spec'
 require_relative 'dynamic_listener_spec'
 require_relative 'filterable_block_spec'
+require_relative 'auto_detect_backend_shared_examples'
 
 require 'support/database_manager'
 
@@ -25,6 +26,7 @@ RSpec.describe "SQLite" do
   include_examples "FilterModel::LikeKeys", db
   include_examples "FilterModel::FilterMapBlockDSL", db
   include_examples "FilterModel::AffixSynonyms", db
+  include_examples "AutoDetectBackend", db
   # Also run the Filterable block DSL shared examples (DB-agnostic)
   include_examples "Filterable::FilterMapBlockDSL"
 end