+++ spec/lib/03_sqlserver_aware_spec.rb
@@ -7,6 +7,7 @@ require_relative 'filter_model/nested_filter_spec'
 require_relative 'filter_model/filter_map_block_spec'
 require_relative 'filter_model/affix_synonyms_spec'
 require_relative 'dynamic_listener_spec'
+require_relative 'auto_detect_backend_shared_examples'
 
 require 'support/database_manager'
 
@@ -24,4 +25,5 @@ RSpec.describe "SQLServer" do
   include_examples "FilterModel::LikeKeys", db
   include_examples "FilterModel::FilterMapBlockDSL", db
   include_examples "FilterModel::AffixSynonyms", db
+  include_examples "AutoDetectBackend", db
 end